Rework ChefSpec::Macros into a part of ChefSpec::API so you only need to include one thing when using the manual override mode.
