+++ lib/chefspec/api.rb
@@ -1,6 +1,7 @@
 module ChefSpec
   module API
     autoload :Core, 'chefspec/api/core'
+    autoload :Described, 'chefspec/api/described'
     autoload :DoNothing, 'chefspec/api/do_nothing'
     autoload :IncludeRecipe, 'chefspec/api/include_recipe'
     autoload :Link, 'chefspec/api/link'
@@ -8,18 +9,21 @@ module ChefSpec
     autoload :Reboot, 'chefspec/api/reboot'
     autoload :RenderFile, 'chefspec/api/render_file'
     autoload :StateAttrs, 'chefspec/api/state_attrs'
+    autoload :Stubs, 'chefspec/api/stubs'
     autoload :Subscriptions, 'chefspec/api/subscriptions'
     autoload :User, 'chefspec/api/user'
 
     def self.included(klass)
       # non-resources
       klass.include(ChefSpec::API::Core)
+      klass.include(ChefSpec::API::Described)
       klass.include(ChefSpec::API::DoNothing)
       klass.include(ChefSpec::API::IncludeRecipe)
       klass.include(ChefSpec::API::DoNothing)
       klass.include(ChefSpec::API::RenderFile)
       klass.include(ChefSpec::API::StateAttrs)
       klass.include(ChefSpec::API::Notifications)
+      klass.include(ChefSpec::API::Stubs)
       klass.include(ChefSpec::API::Subscriptions)
 
       # hacks and sugar for resources that don't follow the normal pattern
